requirements.txt
sanic==23.12.1
sanic-ext==23.12.0
tortoise-orm==0.19.2
pyjwt==2.5.0
jinja2==3.1.3
tortoise-orm[asyncpg]==0.19.2
aerich==0.7.1
aiohttp-requests==0.1.3
openai==1.25.1